requirements.txt
python-decouple==3.1
selenium==3.6.0
SQLAlchemy==1.1.14
